# Use 8 bytes to store int64 components of database keys 