--- a/nodedb_test.go
+++ b/nodedb_test.go
@@ -26,12 +26,13 @@ func makeHashes(b *testing.B, seed int64) [][]byte {
 	b.StopTimer()
 	rnd := rand.NewSource(seed)
 	hashes := make([][]byte, b.N)
+	hashBytes := 8*((hashSize+7)/8)
 	for i := 0; i < b.N; i++ {
-		hashes[i] = make([]byte, 32)
-		for b := 0; b < 32; b += 8 {
+		hashes[i] = make([]byte, hashBytes)
+		for b := 0; b < hashBytes; b += 8 {
 			binary.BigEndian.PutUint64(hashes[i][b:b+8], uint64(rnd.Int63()))
 		}
-		hashes[i] = hashes[i][:20]
+		hashes[i] = hashes[i][:hashSize]
 	}
 	b.StartTimer()
 	return hashes