--- a/nodedb_test.go
+++ b/nodedb_test.go
@@ -1,0 +1,38 @@
+package iavl
+
+import (
+	"encoding/binary"
+	"math/rand"
+	"testing"
+)
+
+func BenchmarkNodeKey(b *testing.B) {
+	ndb := &nodeDB{}
+	hashes := makeHashes(b, 2432325)
+	for i := 0; i < b.N; i++ {
+		ndb.nodeKey(hashes[i])
+	}
+}
+
+func BenchmarkOrphanKey(b *testing.B) {
+	ndb := &nodeDB{}
+	hashes := makeHashes(b, 2432325)
+	for i := 0; i < b.N; i++ {
+		ndb.orphanKey(1234, 1239, hashes[i])
+	}
+}
+
+func makeHashes(b *testing.B, seed int64) [][]byte {
+	b.StopTimer()
+	rnd := rand.NewSource(seed)
+	hashes := make([][]byte, b.N)
+	for i := 0; i < b.N; i++ {
+		hashes[i] = make([]byte, 32)
+		for b := 0; b < 32; b += 8 {
+			binary.BigEndian.PutUint64(hashes[i][b:b+8], uint64(rnd.Int63()))
+		}
+		hashes[i] = hashes[i][:20]
+	}
+	b.StartTimer()
+	return hashes
+}