--- a/testutils_test.go
+++ b/testutils_test.go
@@ -115,8 +115,9 @@ func benchmarkImmutableAvlTreeWithDB(b *testing.B, db db.DB) {
 	b.StopTimer()
 
 	t := NewMutableTree(db, 100000)
+	value := []byte{}
 	for i := 0; i < 1000000; i++ {
-		t.Set(i2b(int(cmn.RandInt32())), nil)
+		t.Set(i2b(int(cmn.RandInt32())), value)
 		if i > 990000 && i%1000 == 999 {
 			t.SaveVersion()
 		}
@@ -124,14 +125,12 @@ func benchmarkImmutableAvlTreeWithDB(b *testing.B, db db.DB) {
 	b.ReportAllocs()
 	t.SaveVersion()
 
-	fmt.Println("ok, starting")
-
 	runtime.GC()
 
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
 		ri := i2b(int(cmn.RandInt32()))
-		t.Set(ri, nil)
+		t.Set(ri, value)
 		t.Remove(ri)
 		if i%100 == 99 {
 			t.SaveVersion()