--- a/tree_test.go
+++ b/tree_test.go
@@ -3,7 +3,6 @@ package iavl
 import (
 	"bytes"
 	"flag"
-	"fmt"
 	"os"
 	"runtime"
 	"testing"
@@ -18,11 +17,11 @@ import (
 
 var testLevelDB bool
 var testFuzzIterations int
-var rand *cmn.Rand
+var random *cmn.Rand
 
 func init() {
-	rand = cmn.NewRand()
-	rand.Seed(0) // for determinism
+	random = cmn.NewRand()
+	random.Seed(0) // for determinism
 	flag.BoolVar(&testLevelDB, "test.leveldb", false, "test leveldb backend")
 	flag.IntVar(&testFuzzIterations, "test.fuzz-iterations", 100000, "number of fuzz testing iterations")
 	flag.Parse()
@@ -55,8 +54,8 @@ func TestVersionedRandomTree(t *testing.T) {
 	// Create a tree of size 1000 with 100 versions.
 	for i := 1; i <= versions; i++ {
 		for j := 0; j < keysPerVersion; j++ {
-			k := []byte(rand.Str(8))
-			v := []byte(rand.Str(8))
+			k := []byte(random.Str(8))
+			v := []byte(random.Str(8))
 			tree.Set(k, v)
 		}
 		tree.SaveVersion()
@@ -97,8 +96,8 @@ func TestVersionedRandomTreeSmallKeys(t *testing.T) {
 	for i := 1; i <= versions; i++ {
 		for j := 0; j < keysPerVersion; j++ {
 			// Keys of size one are likely to be overwritten.
-			k := []byte(rand.Str(1))
-			v := []byte(rand.Str(8))
+			k := []byte(random.Str(1))
+			v := []byte(random.Str(8))
 			tree.Set(k, v)
 			singleVersionTree.Set(k, v)
 		}
@@ -119,7 +118,7 @@ func TestVersionedRandomTreeSmallKeys(t *testing.T) {
 
 	// Try getting random keys.
 	for i := 0; i < keysPerVersion; i++ {
-		_, val := tree.Get([]byte(rand.Str(1)))
+		_, val := tree.Get([]byte(random.Str(1)))
 		require.NotNil(val)
 		require.NotEmpty(val)
 	}
@@ -138,8 +137,8 @@ func TestVersionedRandomTreeSmallKeysRandomDeletes(t *testing.T) {
 	for i := 1; i <= versions; i++ {
 		for j := 0; j < keysPerVersion; j++ {
 			// Keys of size one are likely to be overwritten.
-			k := []byte(rand.Str(1))
-			v := []byte(rand.Str(8))
+			k := []byte(random.Str(1))
+			v := []byte(random.Str(8))
 			tree.Set(k, v)
 			singleVersionTree.Set(k, v)
 		}
@@ -147,7 +146,7 @@ func TestVersionedRandomTreeSmallKeysRandomDeletes(t *testing.T) {
 	}
 	singleVersionTree.SaveVersion()
 
-	for _, i := range rand.Perm(versions - 1) {
+	for _, i := range random.Perm(versions - 1) {
 		tree.DeleteVersion(int64(i + 1))
 	}
 
@@ -160,7 +159,7 @@ func TestVersionedRandomTreeSmallKeysRandomDeletes(t *testing.T) {
 
 	// Try getting random keys.
 	for i := 0; i < keysPerVersion; i++ {
-		_, val := tree.Get([]byte(rand.Str(1)))
+		_, val := tree.Get([]byte(random.Str(1)))
 		require.NotNil(val)
 		require.NotEmpty(val)
 	}
@@ -697,8 +696,8 @@ func TestVersionedCheckpoints(t *testing.T) {
 
 	for i := 1; i <= versions; i++ {
 		for j := 0; j < keysPerVersion; j++ {
-			k := []byte(rand.Str(1))
-			v := []byte(rand.Str(8))
+			k := []byte(random.Str(1))
+			v := []byte(random.Str(8))
 			keys[int64(i)] = append(keys[int64(i)], k)
 			tree.Set(k, v)
 		}
@@ -931,7 +930,7 @@ func TestVersionedTreeEfficiency(t *testing.T) {
 	for i := 1; i <= versions; i++ {
 		for j := 0; j < keysPerVersion; j++ {
 			// Keys of size one are likely to be overwritten.
-			tree.Set([]byte(rand.Str(1)), []byte(rand.Str(8)))
+			tree.Set([]byte(random.Str(1)), []byte(random.Str(8)))
 		}
 		sizeBefore := len(tree.ndb.nodes())
 		tree.SaveVersion()
@@ -1051,7 +1050,7 @@ func TestOrphans(t *testing.T) {
 
 	tree.ndb.traverseOrphans(func(k, v []byte) {
 		var fromVersion, toVersion int64
-		fmt.Sscanf(string(k), orphanKeyFmt, &toVersion, &fromVersion)
+		orphanKeyFormat.Scan(k, &toVersion, &fromVersion)
 		require.Equal(fromVersion, int64(1), "fromVersion should be 1")
 		require.Equal(toVersion, int64(1), "toVersion should be 1")
 	})
@@ -1182,7 +1181,7 @@ func BenchmarkTreeLoadAndDelete(b *testing.B) {
 	tree := NewMutableTree(d, 0)
 	for v := 1; v < numVersions; v++ {
 		for i := 0; i < numKeysPerVersion; i++ {
-			tree.Set([]byte(rand.Str(16)), rand.Bytes(32))
+			tree.Set([]byte(random.Str(16)), random.Bytes(32))
 		}
 		tree.SaveVersion()
 	}
@@ -1203,7 +1202,7 @@ func BenchmarkTreeLoadAndDelete(b *testing.B) {
 			// If we can load quickly into a data-structure that allows for
 			// efficient deletes, we are golden.
 			for v := 0; v < numVersions/10; v++ {
-				version := (rand.Int() % numVersions) + 1
+				version := (random.Int() % numVersions) + 1
 				tree.DeleteVersion(int64(version))
 			}
 		}