@ValarDragon 2018-08-30
This is an incredible performance win!
Does the key being negative matter?