@silasdavis 2018-08-30
@ValarDragon negative gives should be fine - I've added a test.