@silasdavis 2018-09-04
@Liamsi I'll make these changes shortly - thanks.